修复背板shadow设置颜色错误

+++ frameworks/core/components_ng/manager/drag_drop/drag_drop_func_wrapper.cpp
@@ -111,7 +111,7 @@ void DragDropFuncWrapper::ParseShadowInfo(Shadow& shadow, std::unique_ptr<JsonVa
     arkExtraInfoJson->Put("drag_shadow_OffsetX", shadow.GetOffset().GetX());
     arkExtraInfoJson->Put("drag_shadow_OffsetY", shadow.GetOffset().GetY());
     arkExtraInfoJson->Put("shadow_mask", shadow.GetShadowType() == ShadowType::BLUR);
-    int64_t argb = shadow.GetColor().GetValue();
+    int32_t argb = static_cast<int32_t>(shadow.GetColor().GetValue());
     arkExtraInfoJson->Put("drag_shadow_argb", argb);
     int32_t strategy = static_cast<int32_t>(shadow.GetShadowColorStrategy());
     arkExtraInfoJson->Put("shadow_color_strategy", strategy);
